rename channel_id file
